dom/nodes/Node-appendChild-text-and-script-in-style.html

<!doctype html>
<meta charset=utf-8>
<title>Node.appendChild: inserting text and script nodes in a style element</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-node-appendChild">
<link rel=help href="https://dom.spec.whatwg.org/#concept-node-insert">
<link rel=help href="https://github.com/whatwg/dom/issues/575">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<body></body>
<div id="log"></div>
<style id="style"></style>
<script>
var style = document.getElementById("style");

var r1 = new Text("body {}");
var r2 = new Text("body {}");

var script = document.createElement("script");
var scriptRan = false;
script.textContent = `
    assert_not_equals(style.sheet, null, "style sheet is not created");
    assert_equals(style.sheet.cssRules.length, 2, "style sheet does not have two rules");
    scriptRan = true;
`;

var df = document.createDocumentFragment();
df.appendChild(r1);
df.appendChild(script);
df.appendChild(r2);

assert_false(scriptRan, "script ran");
style.appendChild(df);
assert_true(scriptRan, "script did not run");
done();
</script>
